+++ src/components/enterprise-banner/EnterpriseBanner.jsx
@@ -1,6 +1,5 @@
 import React, { useContext } from 'react';
 import { AppContext } from '@edx/frontend-platform/react';
-import OffersAlert from './OffersAlert';
 
 import './styles/EnterpriseBanner.scss';
 
@@ -13,7 +12,6 @@ export default function EnterpriseBanner() {
           {enterpriseConfig.name}
         </h1>
       </div>
-      <OffersAlert />
     </div>
   );
 }
